Moving code to async jobs
In this commit, I installed sidekiq and sidekiq-cron for the
purpose of generating unfollow reports using background jobs. I
implemented a Follower.equals method which is used to compare
followers in the the db against followers fetched from Twitter, to
compute who unfollowed our user. The get_followers method which
called the Twitter API from the dashboard controller is now a User
method. This is used to both load followers into the db (for a
newly registered user) and fetch followers used to create the
unfollow report. Both these functions are now user methods which
get called by the sidekiq job framework. sidekiq-cron gem is used
to schedule the weekly unfollow report generation for each user.
A total field was added to the report model as per project specs.