Bug fix: Duplicate follower relations
The load_followers user method is intended to be called only once.
This design decision has led to a bug where duplicate follower
relations are recorded in the join table when this instance method
is called from console. The method is changed to check for existence
of new follower in current collection. A further fix is required
which adds a unique index to the combination of foreign key columns.